Moved RigRetriever to its own file
